rng: Initial vhost-device-rng implementation
This patch add the initial implementation of the vhost-device-rng
daemon.  It is reponsible of processing requests from a virtual
machine manager (VMM) such as QEMU using the vhost-user protocol.

The implemenation is able to concurrently service multiple VMMs while
making sure the limits of the RNG hardware device (if specified) are
respected.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>