[i2c] separate parsing of strings from device
This separation is needed so we can easily write unit tests for parsing
without needing to create full blown objects. (as a side note, before
this is possible, we also need to get rid of errno, and replace it with
custom Errors, so that we can also write the much needed negative
tests).

This separation is achieved through creating configuration structures
that can be either programatically initialized, or initialized through
parsing the command line parameters.

This commit is still WIP because we also need to make sure that
configuration objects can only be created valid (to reduce some risks
for future extensions where parameters might be passed some other way
rather than yaml). Also, we need to move the check for uniquness of
device addresses in the DeviceConfig instead of the `I2cMap`.

Signed-off-by: Andreea Florescu <fandree@amazon.com>