[i2c] disallow invalid configurations of I2cMap
This helps with simplifying the code by keeping all checks at init time.
In practice it means that invalid objects can no longer be created,
which simplifies the testing.

All objects are created through either new, or `TryFrom`, and they can
only be updated (outside of test functions) via functions that are
checking the validity of the resulting object.

Signed-off-by: Andreea Florescu <fandree@amazon.com>