[i2c] Cleanup tests around socket_count argument
The socket count argument is optional and defaults to 1, but the tests
written for it were more C like and weren't so clean. Use Option to
improve upon that.

While at it, fix socket names as well. Socket name is name of a socket
and shouldn't be used as test-name. Fix the TODOs as well.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>