gpio: Update test's while loop to avoid test failure
The "test_gpio_process_events_multi_success" test currently hangs with
an update to a newer version of Rust. The code here tries to read the
value, locked, for each GPIO one by one. The values are updated in
another thread with the help of write lock.

More discussion around this issue can be found here.

https://github.com/rust-lang/rust/issues/101194

Taking the lock only once fixes it for now.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>