gpio: introduce "mock_gpio" let cfg drive device type
Now we have our plumbing in place we can separate the starting of the
device backend into a new function and call appropriately. As the
config is no longer parsed inside start_backend we can ensure we pass
any errors up rather than blindly unwrapping.

We can drop DeviceUnimplemented as we can't build a version where this
happens.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>