scsi: add support for WRITE SAME(16)
WRITE SAME allows writing a block for a repeated number of times.

Mostly, it can also be used to deallocate parts of the block device
(the fstrim functionality uses this). We do not support that aspect
yet. Instead, we will just stupidly repeat the pattern as many times
as we are told.

A future, smarter implementation could just punch a hole into the
backend instead of filling it with zeros.

Signed-off-by: Erik Schilling <erik.schilling@linaro.org>