scsi: replaced some leftover of vhost-user-scsi
There was some reference left in the documentation and sources to
"vhost-user-scsi" that we had changed during the rebase.
Let's change them to "vhost-device-scsi".

Everything should be safe.

We leave "vhost-user-scsi" in
crates/scsi/src/scsi/emulation/response_data.rs because it looks like
an identifier with some constant size. We will fix in the future.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>