vsock: Add new VMs at runtime
Implement ability to add new VMs at runtime. Removing or modifying
the configuration of an existing VM at runtime is not supported at the
moment.

To use this feature, you must use the `--config` CLI argument to specify
the initial configuration with the `--watch` flag set. The path to the
config file provided must be a symlink to the actual YAML config. To
update the VM configuration at runtime, simply change the symlink to point
to the YAML config file with the new configuration.

Signed-off-by: Priyansh Rathi <techiepriyansh@gmail.com>