gpio: guard libgpiod::Chip with Mutex
The Chip is not `Sync` (as wrongly claimed by the `Sync` marker on
`PhysDevice`). So if we want to share the `PhysDevice` across threads,
we will need to guard it by a `Mutex`.

These calls do not block for very long. So this should not cause any new
deadlock conditions.

Issue: #442
Signed-off-by: Erik Schilling <erik.schilling@linaro.org>