sound/alsa: remove runner thread
This commit removes the need for a runner thread to be created when the
alsa backend is instantiated. This thread is responsible for handling
control requests and communicating with worker threads for starting and
stopping playback/capturing streams. When the virtio-sound daemon gets a
control cmd, a msg is sent over a channel and the runner processes the
request. With this commit, we propose that we handle the msg immediately
without needing to use a thread. The following reasons have led to the
removal of this thread:

* to propagate the error to the device in case an operation fails
* to simplify the backend by removing an extra thread
* to prepare the backend to remove ControlMsg from the header of the
  set_param() and release() methods. Ideally, the audio backend shall
  know nothing about descriptors. By removing the use of a thread to
  handle the control request, notification to guest can happen
  immediately in the device instead of letting the audio backend to handle
  the notifications

Signed-off-by: Matias Ezequiel Vara Larsen <mvaralar@redhat.com>