sound: remove ControlMsg parameter from set_param()
This commit removes the need of the ControlMsg in the set_param()
method. Both backends are able to handle that request immediately so
there is not need to pass the Control msg to notify the guest
afterwards. This task is done by the device. This is the current
mechanism that pw implements and we did not observe evidence to delay
the guest notification. By doing this, this commit removes any
descriptor manipulation from the audio backends.  The commit fixes the
test in pw.

Signed-off-by: Matias Ezequiel Vara Larsen <mvaralar@redhat.com>