vhost-device-media: add virtio-media device
This commit introduces vhost-user-media device.

The virtio-media device first appeared in virtio-spec v1.4 specification[1].
It standardizes how guest applications communicate with media devices on
the host, such as encoders, decoders, and cameras.

The protocol relies on the V4L2 (Video4Linux2) specification to
format the communication between the guest and host, exposing the
host's media devices to the virtualized environment.

The low-level implementation of this backend for the virtio-media
protocol is provided by the device crate from the virtio-media
repository[2].

[1] https://docs.oasis-open.org/virtio/virtio/v1.4/virtio-v1.4.html#x1-82200022
[2] https://github.com/chromeos/virtio-media

Signed-off-by: Albert Esteve <aesteve@redhat.com>