Use a patched version for the `config` dependency
In vhost-device-vsock we use the `config` crate which now has the `yaml-rust`
dependency that is unmaintained as cargo audit is reporting.
See https://github.com/rust-vmm/vhost-device/issues/639

They merged the fix, but for now they don't do a release because they don't
have a maintainer. To get the CI green, let's do this workaround, but before
the next release we have to find a solution.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>