sound/pipewire: destroy pipewire streams not destroyed
The PCM command transition prepare->prepare or set_param->prepare
is allowed by the specification. Pipewire, however, creates multiple
streams which are not destroyed when this type of transition happens.
By using this PR, we can ensure that when the prepare function in the
pipewire backend is called multiple times without the release function,
only one pipewire stream is created.

Signed-off-by: Dorinda Bassey <dbassey@redhat.com>