Add CLI arguments for configuring Rutabaga builder
Add command line arguments for configuring which capsets and features are
enabled when configuring Rutabaga.

Since we now specify the capsets explicitly we can drop the MAX_NUM_CAPSETS
constant and fix the TODO.

If the user selects virglrender the capsets are configured the default capset
configuration is suitable for running virgl/venus guests. If the user select
gfxstream, by default only vulkan is enabled, this is suitable for running
Android with gfxstream.

Signed-off-by: Matej Hrica <mhrica@redhat.com>