scmi: move CLI arg types to lib submodule
Currently, the main CLI struct type is defined in src/main.rs thus
inaccessible when using the crate as a library. Move it to its own
module under src/args.rs

This will be useful in the followup commit that will use src/args.rs
directly to generate manual pages using clap_mangen from the xtask crate
in this workspace.

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>