gpu: extend CI workaround to lib.rs file
Conditionally compile the whole lib.rs file only on GNU environment.
(musl is not supported)
This extends the workaround that was already used in main.rs file to the
public API lib.rs file. This is in preparation of next commits that extend
the API requiring the import of Rutabaga (which doesn't compile in the CI).

Signed-off-by: Matej Hrica <mhrica@redhat.com>